rename package zfmt to compiler/sfmt
The ast and dag packages now live in compiler/ so that seems like the
right home for this package too.
