+++ cmd/super/root/command.go
@@ -13,13 +13,13 @@ import (
 	"github.com/brimdata/super/cli/runtimeflags"
 	"github.com/brimdata/super/compiler"
 	"github.com/brimdata/super/compiler/parser"
+	"github.com/brimdata/super/compiler/sfmt"
 	"github.com/brimdata/super/pkg/charm"
 	"github.com/brimdata/super/pkg/storage"
 	"github.com/brimdata/super/runtime"
 	"github.com/brimdata/super/sbuf"
 	"github.com/brimdata/super/sio"
 	"github.com/brimdata/super/sio/supio"
-	"github.com/brimdata/super/zfmt"
 )
 
 var Super = &charm.Spec{
@@ -137,7 +137,7 @@ func (c *Command) Run(args []string) error {
 		return err
 	}
 	if c.canon {
-		fmt.Println(zfmt.AST(ast.Parsed()))
+		fmt.Println(sfmt.AST(ast.Parsed()))
 		return nil
 	}
 	sctx := super.NewContext()
