+++ compiler/parser/ztests/colon-cast-sfmt.yaml
