+++ compiler/sfmt/ztests/debug.yaml
