+++ compiler/sfmt/ztests/decls.yaml
