+++ compiler/sfmt/ztests/dot.yaml
