+++ compiler/sfmt/ztests/parallel.yaml
