+++ compiler/sfmt/ztests/switch.yaml
