+++ compiler/sfmt/ztests/values-shortcut.yaml
